Write intermediate tractor catalogs

+++ py/legacypipe/detection.py
@@ -181,8 +181,8 @@ def run_sed_matched_filters(SEDs, bands, detmaps, detivs, omit_xy,
     Tnew = fits_table()
     Tnew.ra  = pr
     Tnew.dec = pd
-    Tnew.tx = peakx
-    Tnew.ty = peaky
+    Tnew.tx = peakx.astype(np.float32)
+    Tnew.ty = peaky.astype(np.float32)
     assert(len(peaksn) == len(Tnew))
     assert(len(apsn) == len(Tnew))
     Tnew.peaksn = np.array(peaksn)
